Drop support for Thunderbird versions 68 through 114
There are no plans to add major new features to tbkeys. Older versions
of Thunderbird can continue to use tbkeys 2.2 and won't benefit from
future updates to keep compatibility with newer version of Thunderbird.

This change meets a request from addons.thunderbird.net to drop support
for older versions. It provides extra stability for users of older
versions of Thunderbird. No testing is done with these older versions so
a future update could break tbkeys for them if it was still marked as
compatible.